convert to helper generate aggregate email data
